+++ src/core/index.ts
@@ -1,6 +1,5 @@
 export * from './client';
 export * from './error';
-export * from './http';
 export * from './query';
 export * from './statistic';