+++ src/models/device/paths.ts
@@ -1,4 +1,5 @@
 const REST = 'devices';
+const REST_USER = 'users';
 
 export const DEVICE_PATH = {
     CHECK_USER_ACCESS_DEVICE: `v1/${REST}/check-user-access`,
@@ -10,4 +11,5 @@ export const DEVICE_PATH = {
     REFRESH_DEVICE_TOKEN: `v1/${REST}/refresh-token`,
     STATISTICS: `v1/${REST}/statistics`,
     UPDATE_DEVICE: `v1/${REST}/{id}`,
+    GET_USERS_FROM_DEVICE: `v1/${REST}/{deviceId}/${REST_USER}`,
 };