+++ src/models/device/types.ts
@@ -143,14 +143,9 @@ export interface IGetDevicesParams extends IPageable, ISortable {
     code?: string;
 
     /**
-     * The device name to filter.
+     * The name to filter.
      */
-    deviceName?: string;
-
-    /**
-     * The device token to filter.
-     */
-    deviceToken?: string;
+    name?: string;
 
     /**
      * Filter with is mobile app.
@@ -168,6 +163,11 @@ export interface IGetDevicesParams extends IPageable, ISortable {
      * @remarks enum
      */
     status?: DeviceStatus;
+
+    /**
+     * The keywords to filter.
+     */
+    keywords?: string;
 }
 
 /**
@@ -266,7 +266,34 @@ export interface ICreateDeviceParams {
     simCardInfo?: string | null;
 }
 
-export type IUpdateDeviceParams = ICreateDeviceParams;
+/**
+ * #NOTE: Will update doc and check type in developers.jframework.io in future
+ */
+export interface IUpdateDeviceParams {
+    /**
+     * The name of the device.
+     *
+     * @remarks min: 1
+     */
+    name: string;
+
+    /**
+     * The phone number of the device.
+     */
+    phoneNumber?: string | null;
+
+    /**
+     * The tags of the device.
+     */
+    tags?: string | null;
+
+    /**
+     * The status of the device.
+     *
+     * @remarks enum
+     */
+    status?: DeviceStatus;
+}
 
 /**
  * Represents a request to refresh a mobile device token for Firebase push notifications.