+++ src/models/subscriptionType/constants.ts
@@ -7,3 +7,14 @@ export enum SubscriptionTypeStatus {
     Inactive = 'Inactive',
     Active = 'Active',
 }
+
+/**
+ * Represents the unit of time used for a subscription period.
+ */
+export enum SubscriptionTypeUnit {
+    Hour = 'Hour',
+    Day = 'Day',
+    Week = 'Week',
+    Month = 'Month',
+    Year = 'Year',
+}