+++ src/models/subscriptionType/paths.ts
@@ -1,4 +1,5 @@
 const REST = 'subscription-types';
+const REST_PRICE = 'prices';
 
 export const SUBSCRIPTION_TYPE_PATH = {
     CREATE_SUBSCRIPTION_TYPE: `v1/${REST}`,
@@ -6,4 +7,5 @@ export const SUBSCRIPTION_TYPE_PATH = {
     GET_SUBSCRIPTION_TYPES: `v1/${REST}`,
     GET_SUBSCRIPTION_TYPE: `v1/${REST}/{id}`,
     UPDATE_SUBSCRIPTION_TYPE: `v1/${REST}/{id}`,
+    GET_PRICES_FROM_SUBSCRIPTION_TYPE: `v1/${REST}/{subscriptionTypeId}/${REST_PRICE}`,
 };